refactor: update import paths for UI components to match new structure
- Adjusted import statements for SectionHeader, Spinner, Heading, Input, and TagsInput to reflect their new locations in the folder structure.
- Ensured all affected pages maintain functionality after the changes.
